ext_deps: rename miniz wrapper header to miniz.h.
